Adjust PlatformStackNavigator typescript and styles to SplitNavigators
