+++ src/libs/Navigation/AppNavigator/createResponsiveStackNavigator/CustomRouter.ts
@@ -1,104 +1,29 @@
 import type {CommonActions, RouterConfigOptions, StackActionType, StackNavigationState} from '@react-navigation/native';
 import {findFocusedRoute, StackRouter} from '@react-navigation/native';
 import type {ParamListBase} from '@react-navigation/routers';
+import useActiveWorkspace from '@hooks/useActiveWorkspace';
 import * as Localize from '@libs/Localize';
-import type {PlatformStackRouterOptions} from '@libs/Navigation/PlatformStackNavigation/types';
-import type {RootStackParamList} from '@libs/Navigation/types';
+import isSideModalNavigator from '@libs/Navigation/isSideModalNavigator';
 import {isOnboardingFlowName} from '@libs/NavigationUtils';
 import * as Welcome from '@userActions/Welcome';
 import CONST from '@src/CONST';
+import NAVIGATORS from '@src/NAVIGATORS';
+import SCREENS from '@src/SCREENS';
+import * as GetStateForActionHandlers from './GetStateForActionHandlers';
 import syncBrowserHistory from './syncBrowserHistory';
+import type {CustomRouterAction, CustomRouterActionType, DismissModalActionType, PushActionType, ResponsiveStackNavigatorRouterOptions, SwitchPolicyIdActionType} from './types';
 
-// @TODO: Verify if these functions are still needed
+function isSwitchPolicyIdAction(action: CustomRouterAction): action is SwitchPolicyIdActionType {
+    return action.type === CONST.NAVIGATION.ACTION_TYPE.SWITCH_POLICY_ID;
+}
 
-// function insertRootRoute(state: State<RootStackParamList>, routeToInsert: NavigationPartialRoute) {
-//     const nonModalRoutes = state.routes.filter(
-//         (route) => route.name !== NAVIGATORS.RIGHT_MODAL_NAVIGATOR && route.name !== NAVIGATORS.LEFT_MODAL_NAVIGATOR && route.name !== NAVIGATORS.ONBOARDING_MODAL_NAVIGATOR,
-//     );
-//     const modalRoutes = state.routes.filter(
-//         (route) => route.name === NAVIGATORS.RIGHT_MODAL_NAVIGATOR || route.name === NAVIGATORS.LEFT_MODAL_NAVIGATOR || route.name === NAVIGATORS.ONBOARDING_MODAL_NAVIGATOR,
-//     );
+function isPushAction(action: CustomRouterAction): action is PushActionType {
+    return action.type === CONST.NAVIGATION.ACTION_TYPE.PUSH;
+}
 
-//     // It's safe to modify this state before returning in getRehydratedState.
-
-//     // @ts-expect-error Updating read only property
-//     // noinspection JSConstantReassignment
-//     state.routes = [...nonModalRoutes, routeToInsert, ...modalRoutes]; // eslint-disable-line
-
-//     // @ts-expect-error Updating read only property
-//     // noinspection JSConstantReassignment
-//     state.index = state.routes.length - 1; // eslint-disable-line
-
-//     // @ts-expect-error Updating read only property
-//     // noinspection JSConstantReassignment
-//     state.stale = true; // eslint-disable-line
-// }
-
-// function compareAndAdaptState(state: StackNavigationState<RootStackParamList>) {
-//     // If the state of the last path is not defined the getPathFromState won't work correctly.
-//     if (!state?.routes.at(-1)?.state) {
-//         return;
-//     }
-
-//     // We need to be sure that the bottom tab state is defined.
-//     const topmostBottomTabRoute = getTopmostBottomTabRoute(state);
-//     const isNarrowLayout = getIsNarrowLayout();
-
-//     // This solutions is heuristics and will work for our cases. We may need to improve it in the future if we will have more cases to handle.
-//     if (topmostBottomTabRoute && !isNarrowLayout) {
-//         const fullScreenRoute = state.routes.find((route) => route.name === NAVIGATORS.FULL_SCREEN_NAVIGATOR);
-
-//         // If there is fullScreenRoute we don't need to add anything.
-//         if (fullScreenRoute) {
-//             return;
-//         }
-
-//         // We will generate a template state and compare the current state with it.
-//         // If there is a difference in the screens that should be visible under the overlay, we will add the screen from templateState to the current state.
-//         const pathFromCurrentState = getPathFromState(state, linkingConfig.config);
-//         const {adaptedState: templateState} = getAdaptedStateFromPath(pathFromCurrentState, linkingConfig.config);
-
-//         if (!templateState) {
-//             return;
-//         }
-
-//         const templateFullScreenRoute = templateState.routes.find((route) => route.name === NAVIGATORS.FULL_SCREEN_NAVIGATOR);
-
-//         // If templateFullScreenRoute is defined, and full screen route is not in the state, we need to add it.
-//         if (templateFullScreenRoute) {
-//             insertRootRoute(state, templateFullScreenRoute);
-//             return;
-//         }
-
-//         const topmostCentralPaneRoute = state.routes.filter((route) => isCentralPaneName(route.name)).at(-1);
-//         const templateCentralPaneRoute = templateState.routes.find((route) => isCentralPaneName(route.name));
-
-//         const topmostCentralPaneRouteExtracted = getTopmostCentralPaneRoute(state);
-//         const templateCentralPaneRouteExtracted = getTopmostCentralPaneRoute(templateState as State<RootStackParamList>);
-
-//         // If there is no templateCentralPaneRoute, we don't have anything to add.
-//         if (!templateCentralPaneRoute) {
-//             return;
-//         }
-
-//         // If there is no topmostCentralPaneRoute in the state and template state has one, we need to add it.
-//         if (!topmostCentralPaneRoute) {
-//             insertRootRoute(state, templateCentralPaneRoute);
-//             return;
-//         }
-
-//         // If there is central pane route in state and template state has one, we need to check if they are the same.
-//         if (topmostCentralPaneRouteExtracted && templateCentralPaneRouteExtracted && topmostCentralPaneRouteExtracted.name !== templateCentralPaneRouteExtracted.name) {
-//             // Not every RHP screen has matching central pane defined. In that case we use the REPORT screen as default for initial screen.
-//             // But we don't want to override the central pane for those screens as they may be opened with different central panes under the overlay.
-//             // e.g. i-know-a-teacher may be opened with different central panes under the overlay
-//             if (templateCentralPaneRouteExtracted.name === SCREENS.REPORT) {
-//                 return;
-//             }
-//             insertRootRoute(state, templateCentralPaneRoute);
-//         }
-//     }
-// }
+function isDismissModalAction(action: CustomRouterAction): action is DismissModalActionType {
+    return action.type === CONST.NAVIGATION.ACTION_TYPE.DISMISS_MODAL;
+}
 
 function shouldPreventReset(state: StackNavigationState<ParamListBase>, action: CommonActions.Action | StackActionType) {
     if (action.type !== CONST.NAVIGATION_ACTIONS.RESET || !action?.payload) {
@@ -116,24 +41,62 @@ function shouldPreventReset(state: StackNavigationState<ParamListBase>, action: 
     return false;
 }
 
-function CustomRouter(options: PlatformStackRouterOptions) {
-    const stackRouter = StackRouter(options);
+function isNavigatingToModalFromModal(state: StackNavigationState<ParamListBase>, action: CommonActions.Action | StackActionType) {
+    if (action.type !== CONST.NAVIGATION.ACTION_TYPE.PUSH) {
+        return false;
+    }
 
+    const lastRoute = state.routes.at(-1);
+
+    // If the last route is a side modal navigator and the generated minimal action want's to push a new side modal navigator that means they are different ones.
+    // We want to dismiss the one that is currently on the top.
+    if (isSideModalNavigator(lastRoute?.name) && isSideModalNavigator(action.payload.name)) {
+        return true;
+    }
+    return false;
+}
+
+function CustomRouter(options: ResponsiveStackNavigatorRouterOptions) {
+    const stackRouter = StackRouter(options);
+    const {setActiveWorkspaceID} = useActiveWorkspace();
+
+    // @TODO: Make sure that everything works fine without compareAndAdaptState function. Probably with getMatchingFullScreenRoute.
     return {
         ...stackRouter,
-        getRehydratedState(partialState: StackNavigationState<RootStackParamList>, {routeNames, routeParamList, routeGetIdList}: RouterConfigOptions): StackNavigationState<ParamListBase> {
-            // compareAndAdaptState(partialState);
-            const state = stackRouter.getRehydratedState(partialState, {routeNames, routeParamList, routeGetIdList});
-            return state;
-        },
-        getStateForAction(state: StackNavigationState<ParamListBase>, action: CommonActions.Action | StackActionType, configOptions: RouterConfigOptions) {
+        getStateForAction(state: StackNavigationState<ParamListBase>, action: CustomRouterAction, configOptions: RouterConfigOptions) {
+            if (isSwitchPolicyIdAction(action)) {
+                return GetStateForActionHandlers.handleSwitchPolicyID(state, action, configOptions, stackRouter, setActiveWorkspaceID);
+            }
+
+            if (isDismissModalAction(action)) {
+                return GetStateForActionHandlers.handleDismissModalAction(state, action, configOptions, stackRouter);
+            }
+
+            if (isPushAction(action)) {
+                if (action.payload.name === NAVIGATORS.REPORTS_SPLIT_NAVIGATOR) {
+                    return GetStateForActionHandlers.handlePushReportAction(state, action, configOptions, stackRouter, setActiveWorkspaceID);
+                }
+
+                if (action.payload.name === SCREENS.SEARCH.CENTRAL_PANE) {
+                    return GetStateForActionHandlers.handlePushSearchPageAction(state, action, configOptions, stackRouter, setActiveWorkspaceID);
+                }
+            }
+
+            // Don't let the user navigate back to a non-onboarding screen if they are currently on an onboarding screen and it's not finished.
             if (shouldPreventReset(state, action)) {
                 syncBrowserHistory(state);
                 return state;
             }
+
+            if (isNavigatingToModalFromModal(state, action)) {
+                const modifiedState = {...state, routes: state.routes.slice(0, -1), index: state.index !== 0 ? state.index - 1 : 0};
+                return stackRouter.getStateForAction(modifiedState, action, configOptions);
+            }
+
             return stackRouter.getStateForAction(state, action, configOptions);
         },
     };
 }
 
 export default CustomRouter;
+export type {CustomRouterActionType};
