+++ src/libs/Navigation/AppNavigator/createResponsiveStackNavigator/index.tsx
@@ -5,9 +5,7 @@ import createPlatformStackNavigatorComponent from '@libs/Navigation/PlatformStac
 import defaultPlatformStackScreenOptions from '@libs/Navigation/PlatformStackNavigation/defaultPlatformStackScreenOptions';
 import type {CustomStateHookProps, PlatformStackNavigationEventMap, PlatformStackNavigationOptions, PlatformStackNavigationState} from '@libs/Navigation/PlatformStackNavigation/types';
 import {isFullScreenName} from '@libs/NavigationUtils';
-import type ReactComponentModule from '@src/types/utils/ReactComponentModule';
-
-const loadCustomRouter = require<ReactComponentModule>('@libs/Navigation/AppNavigator/createCustomStackNavigator/CustomRouter').default;
+import CustomRouter from './CustomRouter';
 
 function useCustomRouterState({state}: CustomStateHookProps) {
     const lastSplitIndex = state.routes.findLastIndex((route) => isFullScreenName(route.name));
@@ -17,7 +15,7 @@ function useCustomRouterState({state}: CustomStateHookProps) {
 }
 
 const ResponsiveStackNavigatorComponent = createPlatformStackNavigatorComponent('ResponsiveStackNavigator', {
-    createRouter: loadCustomRouter,
+    createRouter: CustomRouter,
     defaultScreenOptions: defaultPlatformStackScreenOptions,
     useCustomEffects: useNavigationResetOnLayoutChange,
     useCustomState: useCustomRouterState,
