+++ src/libs/Navigation/AppNavigator/createResponsiveStackNavigator/types.ts
